+++ database/migrations/2024_10_15_134547_create_transfers_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::create('transfers', function (Blueprint $table) {
             $table->increments('id');
-            $table->unsignedBigInteger('id_envelope');
+            $table->unsignedInteger('id_envelope');
             $table->string('reference');
             $table->decimal('amount', 10, 2);
             $table->date('date');
