+++ database/migrations/2024_10_15_134931_create_tithes_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::create('tithes', function (Blueprint $table) {
             $table->increments('id');
-            $table->unsignedBigInteger('id_envelope');
+            $table->unsignedInteger('id_envelope');
             $table->decimal('amount', 10, 2);
             $table->timestamps();
